Refactor apiDefs to be more modular - #2124
