+++ src/components/ApiPageReleaseNotes.tsx
@@ -23,12 +23,12 @@ export class ApiPageReleaseNotes extends React.Component<RouteComponentProps<IAp
 
     if (apis.length > 1) {
       const apiCards = apis.map((apiDesc: IApiDescription) => {
-        const { name, shortDescription, urlFragment, vaInternalOnly } = apiDesc;
+        const { description, name, urlFragment, vaInternalOnly } = apiDesc;
         const dashUrlFragment = urlFragment.replace('_', '-')
 
         return (
           <Flag key={name} name={`hosted_apis.${urlFragment}`}>
-            <ApiCard name={name} description={shortDescription} vaInternalOnly={vaInternalOnly}
+            <ApiCard name={name} description={description} vaInternalOnly={vaInternalOnly}
                 url={`/release-notes/${apiCategoryKey}#${dashUrlFragment}`} />
           </Flag>
         );
